Closes #294. Included are all the huds that CortUI came with. They are now separated into folders within /artwork/mods. The in-game 'Style' menu option allows users to choose which /mods/ folder to look for artwork in before trying the normal artwork folder. The cockpit chat pane is resizable within the in-game options menu. This update changes some of the requirements of dialog.mdl, and I included cort's changes to teamscreen.mdl, loadoutpane.mdl, and some others.
I'm about 95% sure that this commit contains all the artwork I've been using.

git-svn-id: file:///media/data/aa/dev/Allegiance-SVN/branch/main@1145 8c327b6c-a73e-8745-8738-111f1e176ae1

+++ src/training/trainingmission.h
@@ -65,7 +65,7 @@ namespace Training
                     std::list<Condition*>   m_KeyConditionList;
                     Condition*              m_pChatCondition;
                     bool                    m_bMusicWasOn;
-                    float                   m_fHUDStyle;
+                    int                     m_fHUDStyle;
                     ObjectID                m_commanderID;
                     SectorID                m_deadSectorID;
                     TrekWindow::ViewMode    m_deadViewMode;
